third_party/WebKit/LayoutTests/typedcssom/stylevalue-objects/parse-parseAll.html

<!doctype html>
<meta charset="utf-8">
<title>CSSStyleValue.parse and CSSStyleValue.parseAll tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-cssstylevalue-parse">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-cssstylevalue-parseall">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/testhelper.js"></script>
<script>
'use strict';

for (const parseMethod of ['parse', 'parseAll']) {
  test(() => {
    assert_throws(new TypeError(), () => CSSStyleValue[parseMethod]('', 'foo'));
  }, 'Calling CSSStyleValue.' + parseMethod + ' with an empty string throws a TypeError');

  test(() => {
    assert_throws(new TypeError(), () => CSSStyleValue[parseMethod]('lemon', 'ade'));
  }, 'Calling CSSStyleValue.' + parseMethod + ' with an unsupported CSS property throws a TypeError');

  test(() => {
    assert_throws(new TypeError(), () => CSSStyleValue[parseMethod]('background', 'red'));
  }, 'Calling CSSStyleValue.' + parseMethod + ' with a CSS shorthand throws a TypeError');

  test(() => {
    assert_throws(new SyntaxError(), () => CSSStyleValue[parseMethod]('width', '10deg'));
  }, 'Calling CSSStyleValue.' + parseMethod + ' with an invalid cssText for the given property throws a SyntaxError');
}

const gValidNonListTests = [
  { property: 'width', value: '10px', expected: CSS.px(10), desc: 'a valid cssText for the given property' },
  { property: 'wIdTh', value: '10px', expected: CSS.px(10), desc: 'a mixed case cssText' },
  { property: '--foo', value: '10%', expected: new CSSUnparsedValue('10%'), desc: 'a custom property' },
];

for (const {property, value, expected, desc} of gValidNonListTests) {
  test(() => {
    const result = CSSStyleValue.parse(property, value);
    assert_style_value_equals(result, expected);
  }, 'Calling CSSStyleValue.parse with ' + desc + ' returns a valid CSSStyleValue');
}

for (const {property, value, expected, desc} of gValidNonListTests) {
  test(() => {
    const result = CSSStyleValue.parseAll(property, value);
    assert_style_value_array_equals(result, [expected]);
  }, 'Calling CSSStyleValue.parseAll with ' + desc + ' returns a single-element list containing a valid CSSStyleValue');
}

test(() => {
  const result = CSSStyleValue.parse('transition-duration', '1s, 2s, 3s');
  assert_style_value_equals(result, CSS.s(1));
}, 'Calling CSSStyleValue.parse with a list-value property returns first list value');

test(() => {
  const result = CSSStyleValue.parseAll('transition-duration', '1s, 2s, 3s');
  assert_equals(result.length, 3);
  for (let i = 0; i < result.length; i++) {
    assert_style_value_equals(result[i], CSS.s(i + 1));
  }
}, 'Calling CSSStyleValue.parseAll with a list-value property returns a sequence of values');

</script>
